requirements.txt
langchain==0.0.208
deeplake
openai
tiktoken
python-dotenv
requests
newspaper3k==0.2.8
streamlit